+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## [Unreleased]
 
+### X.X.X | DD-MM-YYYY
+- add middlewares
+
 ## [Released]
 
 ### 0.0.1 | 08-11-2022
 - add database connection
+- add handlers